Rename log.py to logs.py
