+++ deafrica_waterbodies/cli/continental_run.py
@@ -2,7 +2,7 @@ import math
 import click
 import logging
 
-from .log import logging_setup
+from .logs import logging_setup
 from .io import write_waterbodies_to_file
 from .group_options import MutuallyExclusiveOption
