+++ deafrica_waterbodies/cli/from_vector_file.py
@@ -3,7 +3,7 @@ import click
 import logging
 import geopandas as gpd
 
-from .log import logging_setup
+from .logs import logging_setup
 from .io import write_waterbodies_to_file
 from .group_options import MutuallyExclusiveOption
